Refactor classes out to many more files
