+++ vscode-dotnet-runtime-library/src/Acquisition/DotnetCoreAcquisitionWorker.ts
@@ -49,18 +49,20 @@ import { IAcquisitionWorkerContext } from './IAcquisitionWorkerContext';
 import { IDotnetCoreAcquisitionWorker } from './IDotnetCoreAcquisitionWorker';
 import { IDotnetInstallationContext } from './IDotnetInstallationContext';
 import {
-    GetDotnetInstallInfo,
     InstallRecord,
-    DotnetInstall,
-    InProgressInstallManager,
-    InstallRecordOrStr,
-    installKeyStringToDotnetInstall,
+    InstallRecordOrStr
+} from './InstallRecord';
+import {
+    GetDotnetInstallInfo,
     IsEquivalentInstallationFile,
     InstallToStrings,
-    IsEquivalentInstallation,
-    getVersionFromLegacyInstallKey
-} from './IInstallationRecord';
+    IsEquivalentInstallation
+} from './DotnetInstall';
+import { installKeyStringToDotnetInstall } from '../Utils/InstallKeyUtilities';
+import { getVersionFromLegacyInstallKey } from '../Utils/InstallKeyUtilities';
+import { DotnetInstall } from './DotnetInstall';
 import { InstallationGraveyard } from './InstallationGraveyard';
+import { InProgressInstallManager } from './InProgressInstallManager';
 /* tslint:disable:no-any */
 
 export class DotnetCoreAcquisitionWorker implements IDotnetCoreAcquisitionWorker
@@ -83,7 +85,7 @@ export class DotnetCoreAcquisitionWorker implements IDotnetCoreAcquisitionWorker
         const dotnetExtension = os.platform() === 'win32' ? '.exe' : '';
         this.graveyard = new InstallationGraveyard(context);
         this.dotnetExecutable = `dotnet${dotnetExtension}`;
-        this.acquisitionPromises = new InProgressInstallManager();
+        this.acquisitionPromises = new InProgressInstallManager(this.context);
         // null deliberately allowed to use old behavior below
         this.installingArchitecture = this.context.installingArchitecture === undefined ? os.arch() : this.context.installingArchitecture;
         this.globalResolver = null;
