+++ vscode-dotnet-runtime-library/src/index.ts
@@ -32,8 +32,9 @@ export * from './Utils/VSCodeEnvironment';
 export * from './Utils/IUtilityContext';
 export * from './Utils/WebRequestWorker';
 export * from './Acquisition/DotnetCoreAcquisitionWorker';
+export * from './Acquisition/DotnetInstall';
 export * from './Acquisition/IAcquisitionWorkerContext';
-export * from './Acquisition/IInstallationRecord';
+export * from './Acquisition/InstallRecord';
 export * from './Acquisition/AcquisitionInvoker';
 export * from './Acquisition/DotnetCoreDependencyInstaller'
 export * from './Acquisition/LinuxVersionResolver';
